Don't cross compile GetAssemblyVersion
Previously we cross-compiled the GetAssemblyVersion method to avoid
a dependency on MetadataReader on desktop. This had the downside of
causing first chance exceptions when run on desktop.

This has the tradeoff of introducing a Reflection.Metadata dependency on
desktop which will increase the layout by ~700KB (for SRM and SCI).
